Enforce integer flow actions across env

+++ tests/test_network.py
@@ -31,9 +31,9 @@ def test_flow_consistency(sample_files):
     scenario.load_network_from_json(network_path)
     scenario.load_demand_from_json(demand_path)
 
-    assert scenario.check_flow_consistency(np.array([10, 0], dtype=float)) is True
-    assert scenario.check_flow_consistency(np.array([5, 0], dtype=float)) is False
-    assert scenario.check_flow_consistency(np.array([10, 5], dtype=float)) is False
+    assert scenario.check_flow_consistency(np.array([10, 0], dtype=np.int32)) is True
+    assert scenario.check_flow_consistency(np.array([5, 0], dtype=np.int32)) is False
+    assert scenario.check_flow_consistency(np.array([10, 5], dtype=np.int32)) is False
 
 
 def test_system_cost_requires_consistent_flow(sample_files):
@@ -44,9 +44,9 @@ def test_system_cost_requires_consistent_flow(sample_files):
     scenario.load_network_from_json(network_path)
     scenario.load_demand_from_json(demand_path)
 
-    scenario.check_flow_consistency(np.array([10, 0], dtype=float))
-    total_cost = scenario.calculate_system_cost(np.array([10, 0], dtype=float))
+    scenario.check_flow_consistency(np.array([10, 0], dtype=np.int32))
+    total_cost = scenario.calculate_system_cost(np.array([10, 0], dtype=np.int32))
     assert total_cost == 690.0 # f(10) = 60*(1 + 0.15*(10/10)^4) = 69 per unit flow, total 690
 
     with pytest.raises(ValueError):
-        scenario.calculate_system_cost(np.array([0, 0], dtype=float))
+        scenario.calculate_system_cost(np.array([0, 0], dtype=np.int32))
